Added alerts to search and feedback
I got the alerts to work on all the search pages except Jonnys because
his is way different, so we’ll need him to look at that. I also got it
to only send feedback if both fields have data. It will show an alert
if there aren’t any.